refactor: extract sections into components
